+++ index.js
@@ -44,12 +44,10 @@ import NavigationPage from './components/NavigationPage/NavigationPage';
 
 import KeyboardSpace from './components/KeyboardSpace/KeyboardSpace';
 import BackHandler from './utils/BackHandler';
-import NavigationHelper1 from './navigation/NavigationHelper';
 
 
 var Teaset = {
   Theme,
-  NavigationHelper: NavigationHelper1,
 
   Label,
   Button,